+++ cherrypic-api/src/main/java/org/cherrypic/domain/payment/dto/request/PaymentReadyRequest.java
@@ -1,4 +1,4 @@
-package org.cherrypic.domain.payment.dto;
+package org.cherrypic.domain.payment.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
