+++ cherrypic-api/src/main/java/org/cherrypic/domain/payment/service/PaymentService.java
@@ -1,7 +1,7 @@
 package org.cherrypic.domain.payment.service;
 
-import org.cherrypic.domain.payment.dto.PaymentReadyRequest;
-import org.cherrypic.domain.payment.dto.PaymentReadyResponse;
+import org.cherrypic.domain.payment.dto.request.PaymentReadyRequest;
+import org.cherrypic.domain.payment.dto.response.PaymentReadyResponse;
 
 public interface PaymentService {
     PaymentReadyResponse preparePayment(PaymentReadyRequest request);
