+++ cherrypic-api/src/main/java/org/cherrypic/domain/payment/service/PaymentServiceImpl.java
@@ -11,8 +11,8 @@ import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.cherrypic.album.enums.AlbumPlan;
-import org.cherrypic.domain.payment.dto.PaymentReadyRequest;
-import org.cherrypic.domain.payment.dto.PaymentReadyResponse;
+import org.cherrypic.domain.payment.dto.request.PaymentReadyRequest;
+import org.cherrypic.domain.payment.dto.response.PaymentReadyResponse;
 import org.cherrypic.domain.payment.exception.PaymentErrorCode;
 import org.cherrypic.domain.payment.exception.PaymentException;
 import org.cherrypic.global.util.MemberUtil;
